disabled received contact photo panel

+++ www/fns/ContactPhotos/delete.php
@@ -4,12 +4,25 @@ namespace ContactPhotos;
 
 function delete ($mysqli, $id) {
 
-    $sql = "delete from contact_photos where id = $id";
-    $mysqli->query($sql) || trigger_error($mysqli->error);
+    $sql = "select * from contact_photos where id = $id";
+    $contactPhoto = mysqli_single_object($mysqli, $sql);
 
-    include_once __DIR__.'/path.php';
-    $path = path($id);
+    if ($contactPhoto) {
+        if ($contactPhoto->num_refs == 1) {
 
-    if (is_file($path)) unlink($path);
+            $sql = "delete from contact_photos where id = $id";
+            $mysqli->query($sql) || trigger_error($mysqli->error);
+
+            include_once __DIR__.'/path.php';
+            $path = path($id);
+
+            if (is_file($path)) unlink($path);
+
+        } else {
+            $sql = 'update contact_photos set num_refs = num_refs - 1'
+                ." where id = $id";
+            $mysqli->query($sql) || trigger_error($mysqli->error);
+        }
+    }
 
 }
